@@ -1,24 +1,24 @@
 name: flutter_reactive_ble_example
 description: Demonstrates how to use the flutter_reactive_ble plugin.
-version: 5.0.3
+version: 5.2.0
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_lints: ^1.0.4
-  flutter_reactive_ble: ^5.0.3
+  flutter_reactive_ble: ^5.2.0
   functional_data: ^1.0.0
   intl: ^0.17.0
 
   provider: ^6.0.1
 
 dev_dependencies:
-  build_runner: ^2.1.4
+  build_runner: ^2.3.3
   dependency_validator: ^3.1.0
   flutter_test:
     sdk: flutter
@@ -29,6 +29,8 @@ dependency_overrides:
     path: ../packages/flutter_reactive_ble
   reactive_ble_mobile:
     path: ../packages/reactive_ble_mobile
+  reactive_ble_platform_interface:
+    path: ../packages/reactive_ble_platform_interface
 
 flutter:
   uses-material-design: true
