@@ -1,10 +1,10 @@
 name: flutter_reactive_ble
 description: Reactive Bluetooth Low Energy (BLE) plugin that can communicate with multiple devices
-version: 5.0.3
+version: 5.2.0
 homepage: https://github.com/PhilipsHue/flutter_reactive_ble
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
   flutter: ">=2.0.0"
 
 flutter:
@@ -21,10 +21,17 @@ dependencies:
     sdk: flutter
   functional_data: ^1.0.0
   meta: ^1.3.0
-  reactive_ble_mobile: ^5.0.3
-  reactive_ble_platform_interface: ^5.0.3
+  reactive_ble_mobile: ^5.2.0
+  reactive_ble_platform_interface: ^5.2.0
+
+dependency_overrides:
+  reactive_ble_mobile:
+    path: ../reactive_ble_mobile
+  reactive_ble_platform_interface:
+    path: ../reactive_ble_platform_interface
+
 dev_dependencies:
-  build_runner: ^2.1.2
+  build_runner: ^2.3.3
   flutter_lints: ^1.0.4
   flutter_test:
     sdk: flutter
