@@ -1,10 +1,10 @@
 name: reactive_ble_platform_interface
 description: Platform interface for the flutter_reactive_ble_project 
-version: 5.0.3
+version: 5.2.0
 homepage: https://github.com/PhilipsHue/flutter_reactive_ble
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
   flutter: ">=2.0.0"
 
 dependencies:
@@ -17,7 +17,7 @@ dependencies:
   plugin_platform_interface: ^2.0.1
 
 dev_dependencies:
-  build_runner: ^2.1.2
+  build_runner: ^2.3.3
   flutter_lints: ^1.0.3
   flutter_test:
     sdk: flutter
