@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:flutter_reactive_ble/src/connected_device_operation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
@@ -26,9 +25,11 @@ void main() {
 
       setUp(() {
         valueUpdate = CharacteristicValue(
-          characteristic: QualifiedCharacteristic(
+          characteristic: CharacteristicInstance(
             characteristicId: Uuid.parse('FEFF'),
+            characteristicInstanceId: "11",
             serviceId: Uuid.parse('FEFF'),
+            serviceInstanceId: "101",
             deviceId: '123',
           ),
           result: const Result.success([1]),
@@ -47,30 +48,36 @@ void main() {
     });
 
     group('Read characteristic', () {
-      late QualifiedCharacteristic charDevice;
-      QualifiedCharacteristic charOtherSameDevice;
-      QualifiedCharacteristic charOtherDevice;
+      late CharacteristicInstance charDevice;
+      CharacteristicInstance charOtherSameDevice;
+      CharacteristicInstance charOtherDevice;
       CharacteristicValue? valueUpdate;
       CharacteristicValue? valueUpdateOtherDevice;
       CharacteristicValue? valueUpdateSameDeviceOtherChar;
       List<int>? result;
 
       setUp(() {
-        charDevice = QualifiedCharacteristic(
+        charDevice = CharacteristicInstance(
           characteristicId: Uuid.parse('FEFF'),
+          characteristicInstanceId: "11",
           serviceId: Uuid.parse('FEFF'),
+          serviceInstanceId: "101",
           deviceId: '123',
         );
 
-        charOtherSameDevice = QualifiedCharacteristic(
+        charOtherSameDevice = CharacteristicInstance(
           characteristicId: Uuid.parse('FEFF'),
+          characteristicInstanceId: "11",
           serviceId: Uuid.parse('FAFF'),
+          serviceInstanceId: "101",
           deviceId: '123',
         );
 
-        charOtherDevice = QualifiedCharacteristic(
+        charOtherDevice = CharacteristicInstance(
           characteristicId: Uuid.parse('FEFF'),
+          characteristicInstanceId: "11",
           serviceId: Uuid.parse('FEFF'),
+          serviceInstanceId: "101",
           deviceId: '456',
         );
 
@@ -132,14 +139,16 @@ void main() {
     });
 
     group('Write characteristic', () {
-      late QualifiedCharacteristic characteristic;
+      late CharacteristicInstance characteristic;
       WriteCharacteristicInfo info;
       const value = [1, 0];
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: Uuid.parse('FEFF'),
+          characteristicInstanceId: "11",
           serviceId: Uuid.parse('FEFF'),
+          serviceInstanceId: "101",
           deviceId: '123',
         );
       });
@@ -246,9 +255,9 @@ void main() {
       });
 
       group('Subscribe to characteristic', () {
-        late QualifiedCharacteristic charDevice;
-        QualifiedCharacteristic charOtherSameDevice;
-        QualifiedCharacteristic charOtherDevice;
+        late CharacteristicInstance charDevice;
+        CharacteristicInstance charOtherSameDevice;
+        CharacteristicInstance charOtherDevice;
         late CharacteristicValue valueUpdate1;
         late CharacteristicValue valueUpdate2;
         late CharacteristicValue valueUpdateOtherDevice;
@@ -259,21 +268,27 @@ void main() {
         setUp(() {
           terminateCompleter = Completer();
 
-          charDevice = QualifiedCharacteristic(
+          charDevice = CharacteristicInstance(
             characteristicId: Uuid.parse('FEFF'),
+            characteristicInstanceId: "11",
             serviceId: Uuid.parse('FEFF'),
+            serviceInstanceId: "101",
             deviceId: '123',
           );
 
-          charOtherSameDevice = QualifiedCharacteristic(
+          charOtherSameDevice = CharacteristicInstance(
             characteristicId: Uuid.parse('FEFF'),
+            characteristicInstanceId: "11",
             serviceId: Uuid.parse('FAFF'),
+            serviceInstanceId: "101",
             deviceId: '123',
           );
 
-          charOtherDevice = QualifiedCharacteristic(
+          charOtherDevice = CharacteristicInstance(
             characteristicId: Uuid.parse('FEFF'),
+            characteristicInstanceId: "11",
             serviceId: Uuid.parse('FEFF'),
+            serviceInstanceId: "101",
             deviceId: '456',
           );
 
