@@ -8,6 +8,7 @@ import 'package:flutter_reactive_ble/src/device_scanner.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:reactive_ble_platform_interface/reactive_ble_platform_interface.dart';
 
 import 'reactive_ble_test.mocks.dart';
 
@@ -47,8 +48,8 @@ void main() {
         (_) => Future.value(),
       );
 
-      when(_blePlatform.bleStatusStream).thenAnswer(
-          (realInvocation) => _bleStatusController.stream.asBroadcastStream());
+      when(_blePlatform.bleStatusStream)
+          .thenAnswer((realInvocation) => _bleStatusController.stream.asBroadcastStream());
 
       _sut = FlutterReactiveBle.witDependencies(
         reactiveBlePlatform: _blePlatform,
@@ -97,17 +98,21 @@ void main() {
     });
 
     group('CharacteristicValueStream', () {
-      const characteristic = Result<List<int>,
-          GenericFailure<CharacteristicValueUpdateError>>.success(
+      const characteristic = Result<List<int>, GenericFailure<CharacteristicValueUpdateError>>.success(
         [1],
       );
-      final charValue = CharacteristicValue(
-          characteristic: _createChar(), result: characteristic);
+      final characteristicInstance = CharacteristicInstance(
+        deviceId: "1",
+        characteristicId: Uuid.parse('FEFF'),
+        characteristicInstanceId: "11",
+        serviceId: Uuid.parse('F0FF'),
+        serviceInstanceId: "101",
+      );
+      final charValue = CharacteristicValue(characteristic: characteristicInstance, result: characteristic);
       Stream<CharacteristicValue>? charValueStream;
 
       setUp(() {
-        when(_deviceOperation.characteristicValueStream)
-            .thenAnswer((_) => Stream.fromIterable([charValue]));
+        when(_deviceOperation.characteristicValueStream).thenAnswer((_) => Stream.fromIterable([charValue]));
         charValueStream = _sut.characteristicValueStream;
       });
 
@@ -124,24 +129,63 @@ void main() {
     group('Deinitialize', () {
       setUp(() async {
         when(_blePlatform.deinitialize()).thenAnswer((_) async => 1);
-        await _sut.deinitialize();
       });
 
-      test('It executes deinitialize succesfull', () {
-        expect(true, true);
+      test('It executes deinitialize successful', () async {
+        await _sut.deinitialize();
       });
     });
 
     group('Read characteristic', () {
-      QualifiedCharacteristic characteristic;
+      late QualifiedCharacteristic characteristic;
       List<int>? result;
 
       setUp(() async {
         characteristic = _createChar();
-        when(_deviceOperation.readCharacteristic(any))
-            .thenAnswer((_) async => [1]);
+        when(_deviceOperation.getDiscoverServices(characteristic.deviceId)).thenAnswer((_) async => [
+              DiscoveredService(
+                serviceId: characteristic.serviceId,
+                serviceInstanceId: "11",
+                characteristicIds: [characteristic.characteristicId],
+                includedServices: [],
+                characteristics: [
+                  DiscoveredCharacteristic(
+                    characteristicId: Uuid.parse("1234"),
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                  DiscoveredCharacteristic(
+                    characteristicId: characteristic.characteristicId,
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                ],
+              )
+            ]);
+        when(_deviceOperation.readCharacteristic(any)).thenAnswer((_) async => [1]);
+        when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => const Stream.empty());
 
         result = await _sut.readCharacteristic(characteristic);
+      });
+
+      test("It reads the correct characteristic", () {
+        verify(_deviceOperation.readCharacteristic(CharacteristicInstance(
+          characteristicId: characteristic.characteristicId,
+          characteristicInstanceId: "101",
+          serviceId: characteristic.serviceId,
+          serviceInstanceId: "11",
+          deviceId: characteristic.deviceId,
+        ))).called(1);
       });
 
       test('It returns correct value', () {
@@ -151,7 +195,7 @@ void main() {
 
     group('Write characteristic with response', () {
       const value = [2];
-      QualifiedCharacteristic characteristic;
+      late QualifiedCharacteristic characteristic;
 
       setUp(() async {
         characteristic = _createChar();
@@ -160,19 +204,58 @@ void main() {
           any,
           value: anyNamed('value'),
         )).thenAnswer((_) async => [0]);
+        when(_deviceOperation.getDiscoverServices(characteristic.deviceId)).thenAnswer((_) async => [
+              DiscoveredService(
+                serviceId: characteristic.serviceId,
+                serviceInstanceId: "11",
+                characteristicIds: [characteristic.characteristicId],
+                includedServices: [],
+                characteristics: [
+                  DiscoveredCharacteristic(
+                    characteristicId: Uuid.parse("1234"),
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                  DiscoveredCharacteristic(
+                    characteristicId: characteristic.characteristicId,
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                ],
+              )
+            ]);
+        when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => const Stream.empty());
 
-        await _sut.writeCharacteristicWithResponse(characteristic,
-            value: value);
+        await _sut.writeCharacteristicWithResponse(characteristic, value: value);
       });
 
-      test('It completes operation without errors', () {
-        expect(true, true);
+      test("It write to the correct characteristic", () {
+        verify(_deviceOperation.writeCharacteristicWithResponse(
+          CharacteristicInstance(
+            characteristicId: characteristic.characteristicId,
+            characteristicInstanceId: "101",
+            serviceId: characteristic.serviceId,
+            serviceInstanceId: "11",
+            deviceId: characteristic.deviceId,
+          ),
+          value: [2],
+        )).called(1);
       });
     });
 
     group('Write characteristic without response', () {
       const value = [2];
-      QualifiedCharacteristic characteristic;
+      late QualifiedCharacteristic characteristic;
 
       setUp(() async {
         characteristic = _createChar();
@@ -181,6 +264,37 @@ void main() {
           any,
           value: anyNamed('value'),
         )).thenAnswer((_) async => [0]);
+        when(_deviceOperation.getDiscoverServices(characteristic.deviceId)).thenAnswer((_) async => [
+              DiscoveredService(
+                serviceId: characteristic.serviceId,
+                serviceInstanceId: "11",
+                characteristicIds: [characteristic.characteristicId],
+                includedServices: [],
+                characteristics: [
+                  DiscoveredCharacteristic(
+                    characteristicId: Uuid.parse("1234"),
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                  DiscoveredCharacteristic(
+                    characteristicId: characteristic.characteristicId,
+                    characteristicInstanceId: "101",
+                    serviceId: characteristic.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                ],
+              )
+            ]);
+        when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => const Stream.empty());
 
         await _sut.writeCharacteristicWithoutResponse(
           characteristic,
@@ -188,8 +302,17 @@ void main() {
         );
       });
 
-      test('It completes operation without errors', () {
-        expect(true, true);
+      test("It write to the correct characteristic", () {
+        verify(_deviceOperation.writeCharacteristicWithoutResponse(
+          CharacteristicInstance(
+            characteristicId: characteristic.characteristicId,
+            characteristicInstanceId: "101",
+            serviceId: characteristic.serviceId,
+            serviceInstanceId: "11",
+            deviceId: characteristic.deviceId,
+          ),
+          value: [2],
+        )).called(1);
       });
     });
 
@@ -199,8 +322,7 @@ void main() {
       int? result;
 
       setUp(() async {
-        when(_deviceOperation.requestMtu(any, any))
-            .thenAnswer((_) async => mtu);
+        when(_deviceOperation.requestMtu(any, any)).thenAnswer((_) async => mtu);
 
         result = await _sut.requestMtu(deviceId: deviceId, mtu: mtu);
       });
@@ -215,10 +337,8 @@ void main() {
       const priority = ConnectionPriority.highPerformance;
 
       setUp(() async {
-        when(_deviceOperation.requestConnectionPriority(any, any))
-            .thenAnswer((_) async => 0);
-        await _sut.requestConnectionPriority(
-            deviceId: deviceId, priority: priority);
+        when(_deviceOperation.requestConnectionPriority(any, any)).thenAnswer((_) async => 0);
+        await _sut.requestConnectionPriority(deviceId: deviceId, priority: priority);
       });
 
       test('It completes operation without errors', () {
@@ -236,6 +356,7 @@ void main() {
         manufacturerData: Uint8List.fromList([0]),
         name: 'test',
         rssi: -39,
+        connectable: Connectable.unknown,
         serviceData: const {},
         serviceUuids: const [],
       );
@@ -245,8 +366,7 @@ void main() {
         when(_deviceScanner.scanForDevices(
           withServices: anyNamed('withServices'),
           scanMode: anyNamed('scanMode'),
-          requireLocationServicesEnabled:
-              anyNamed('requireLocationServicesEnabled'),
+          requireLocationServicesEnabled: anyNamed('requireLocationServicesEnabled'),
         )).thenAnswer((_) => Stream.fromIterable([device]));
 
         deviceStream = _sut.scanForDevices(
@@ -283,8 +403,7 @@ void main() {
       setUp(() {
         when(_deviceConnector.connect(
                 id: anyNamed('id'),
-                servicesWithCharacteristicsToDiscover:
-                    anyNamed('servicesWithCharacteristicsToDiscover'),
+                servicesWithCharacteristicsToDiscover: anyNamed('servicesWithCharacteristicsToDiscover'),
                 connectionTimeout: anyNamed('connectionTimeout')))
             .thenAnswer((realInvocation) => Stream.fromIterable([update]));
 
@@ -324,8 +443,7 @@ void main() {
       setUp(() {
         when(_deviceConnector.connectToAdvertisingDevice(
           id: anyNamed('id'),
-          servicesWithCharacteristicsToDiscover:
-              anyNamed('servicesWithCharacteristicsToDiscover'),
+          servicesWithCharacteristicsToDiscover: anyNamed('servicesWithCharacteristicsToDiscover'),
           connectionTimeout: anyNamed('connectionTimeout'),
           prescanDuration: anyNamed('prescanDuration'),
           withServices: anyNamed('withServices'),
@@ -358,8 +476,7 @@ void main() {
       );
 
       setUp(() async {
-        when(_blePlatform.clearGattCache('123'))
-            .thenAnswer((_) async => result);
+        when(_blePlatform.clearGattCache('123')).thenAnswer((_) async => result);
 
         await _sut.clearGattCache(deviceId);
       });
@@ -409,16 +526,44 @@ void main() {
 
       setUp(() {
         char = _createChar();
-        when(_deviceConnector.deviceConnectionStateUpdateStream)
-            .thenAnswer((_) => Stream.fromIterable([update]));
+        when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => Stream.fromIterable([update]));
+        when(_deviceOperation.getDiscoverServices(char.deviceId)).thenAnswer((_) async => [
+              DiscoveredService(
+                serviceId: char.serviceId,
+                serviceInstanceId: "11",
+                characteristicIds: [char.characteristicId],
+                includedServices: [],
+                characteristics: [
+                  DiscoveredCharacteristic(
+                    characteristicId: Uuid.parse("1234"),
+                    characteristicInstanceId: "101",
+                    serviceId: char.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                  DiscoveredCharacteristic(
+                    characteristicId: char.characteristicId,
+                    characteristicInstanceId: "101",
+                    serviceId: char.serviceId,
+                    isReadable: true,
+                    isWritableWithResponse: true,
+                    isWritableWithoutResponse: true,
+                    isNotifiable: true,
+                    isIndicatable: true,
+                  ),
+                ],
+              )
+            ]);
 
         valueStream = Stream.fromIterable([
           [1],
           [2]
         ]);
 
-        when(_deviceOperation.subscribeToCharacteristic(any, any))
-            .thenAnswer((_) => valueStream);
+        when(_deviceOperation.subscribeToCharacteristic(any, any)).thenAnswer((_) => valueStream);
 
         resultStream = _sut.subscribeToCharacteristic(char);
       });
@@ -443,12 +588,171 @@ void main() {
         const result = <DiscoveredService>[];
 
         setUp(() {
-          when(_deviceOperation.discoverServices(any))
-              .thenAnswer((_) async => result);
+          when(_deviceOperation.discoverServices(any)).thenAnswer((_) async => result);
         });
 
         test('It returns result', () async {
+          // ignore: deprecated_member_use_from_same_package
           expect(await _sut.discoverServices(deviceId), <DiscoveredService>[]);
+        });
+      });
+    });
+
+    group('Discover all services', () {
+      const deviceId = '123';
+
+      group('When operation is successful', () {
+        const result = <DiscoveredService>[];
+
+        setUp(() {
+          when(_deviceOperation.discoverServices(any)).thenAnswer((_) async => result);
+        });
+
+        test('It succeeds', () async {
+          await _sut.discoverAllServices(deviceId);
+          verify(_deviceOperation.discoverServices(deviceId)).called(1);
+        });
+      });
+    });
+
+    group("getDiscoveredServices", () {
+      const deviceId = "123";
+
+      group("multiple characteristics with same id in single service", () {
+        setUp(() {
+          when(_deviceOperation.getDiscoverServices(deviceId)).thenAnswer((_) async => [
+                DiscoveredService(
+                  serviceId: Uuid.parse("ff01"),
+                  serviceInstanceId: "11",
+                  characteristicIds: [Uuid.parse("aa01")],
+                  includedServices: [],
+                  characteristics: [
+                    DiscoveredCharacteristic(
+                      characteristicId: Uuid.parse("aa01"),
+                      characteristicInstanceId: "101",
+                      serviceId: Uuid.parse("ff01"),
+                      isReadable: true,
+                      isWritableWithResponse: true,
+                      isWritableWithoutResponse: true,
+                      isNotifiable: true,
+                      isIndicatable: true,
+                    ),
+                    DiscoveredCharacteristic(
+                      characteristicId: Uuid.parse("aa01"),
+                      characteristicInstanceId: "102",
+                      serviceId: Uuid.parse("ff01"),
+                      isReadable: true,
+                      isWritableWithResponse: true,
+                      isWritableWithoutResponse: true,
+                      isNotifiable: true,
+                      isIndicatable: true,
+                    ),
+                  ],
+                )
+              ]);
+          when(_deviceOperation.readCharacteristic(any)).thenAnswer((_) async => [42]);
+          when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => const Stream.empty());
+        });
+
+        test("reading first instance of characteristic", () async {
+          final services = await _sut.getDiscoveredServices("123");
+
+          expect(await services.single.characteristics.first.read(), [42]);
+
+          verify(_deviceOperation.readCharacteristic(CharacteristicInstance(
+            characteristicId: Uuid.parse("aa01"),
+            characteristicInstanceId: "101",
+            serviceId: Uuid.parse("ff01"),
+            serviceInstanceId: "11",
+            deviceId: "123",
+          )));
+        });
+
+        test("reading second instance of characteristic", () async {
+          final services = await _sut.getDiscoveredServices("123");
+
+          expect(await services.single.characteristics[1].read(), [42]);
+
+          verify(_deviceOperation.readCharacteristic(CharacteristicInstance(
+            characteristicId: Uuid.parse("aa01"),
+            characteristicInstanceId: "102",
+            serviceId: Uuid.parse("ff01"),
+            serviceInstanceId: "11",
+            deviceId: "123",
+          )));
+        });
+      });
+
+      group("multiple characteristics with same id in different service", () {
+        setUp(() {
+          when(_deviceOperation.getDiscoverServices(deviceId)).thenAnswer((_) async => [
+            DiscoveredService(
+              serviceId: Uuid.parse("ff01"),
+              serviceInstanceId: "11",
+              characteristicIds: [Uuid.parse("aa01")],
+              includedServices: [],
+              characteristics: [
+                DiscoveredCharacteristic(
+                  characteristicId: Uuid.parse("aa01"),
+                  characteristicInstanceId: "101",
+                  serviceId: Uuid.parse("ff01"),
+                  isReadable: true,
+                  isWritableWithResponse: true,
+                  isWritableWithoutResponse: true,
+                  isNotifiable: true,
+                  isIndicatable: true,
+                ),
+              ],
+            ),
+            DiscoveredService(
+              serviceId: Uuid.parse("ff01"),
+              serviceInstanceId: "12",
+              characteristicIds: [Uuid.parse("aa01")],
+              includedServices: [],
+              characteristics: [
+                DiscoveredCharacteristic(
+                  characteristicId: Uuid.parse("aa01"),
+                  characteristicInstanceId: "101",
+                  serviceId: Uuid.parse("ff01"),
+                  isReadable: true,
+                  isWritableWithResponse: true,
+                  isWritableWithoutResponse: true,
+                  isNotifiable: true,
+                  isIndicatable: true,
+                ),
+              ],
+            ),
+          ]);
+          when(_deviceConnector.deviceConnectionStateUpdateStream).thenAnswer((_) => const Stream.empty());
+          when(_deviceOperation.readCharacteristic(any)).thenAnswer((_) async => [42]);
+        });
+
+        test("reading first instance of characteristic", () async {
+          final services = await _sut.getDiscoveredServices("123");
+
+          expect(await services.first.characteristics.single.read(), [42]);
+
+          verify(_deviceOperation.readCharacteristic(CharacteristicInstance(
+            characteristicId: Uuid.parse("aa01"),
+            characteristicInstanceId: "101",
+            serviceId: Uuid.parse("ff01"),
+            serviceInstanceId: "11",
+            deviceId: "123",
+          )));
+        });
+
+        test("reading second instance of characteristic", () async {
+          final services = await _sut.getDiscoveredServices("123");
+
+          expect(await services[1].characteristics.single.read(), [42]);
+
+          verify(_deviceOperation.readCharacteristic(CharacteristicInstance(
+            characteristicId: Uuid.parse("aa01"),
+            characteristicInstanceId: "101",
+            serviceId: Uuid.parse("ff01"),
+            serviceInstanceId: "12",
+            deviceId: "123",
+          )));
         });
       });
     });
