@@ -10,18 +10,21 @@ abstract class $DeviceInteractionViewModel {
   const $DeviceInteractionViewModel();
 
   String get deviceId;
+  Connectable get connectableStatus;
   DeviceConnectionState get connectionStatus;
   BleDeviceConnector get deviceConnector;
-  Future<List<DiscoveredService>> Function() get discoverServices;
+  Future<List<Service>> Function() get discoverServices;
 
   DeviceInteractionViewModel copyWith({
     String? deviceId,
+    Connectable? connectableStatus,
     DeviceConnectionState? connectionStatus,
     BleDeviceConnector? deviceConnector,
-    Future<List<DiscoveredService>> Function()? discoverServices,
+    Future<List<Service>> Function()? discoverServices,
   }) =>
       DeviceInteractionViewModel(
         deviceId: deviceId ?? this.deviceId,
+        connectableStatus: connectableStatus ?? this.connectableStatus,
         connectionStatus: connectionStatus ?? this.connectionStatus,
         deviceConnector: deviceConnector ?? this.deviceConnector,
         discoverServices: discoverServices ?? this.discoverServices,
@@ -31,6 +34,7 @@ abstract class $DeviceInteractionViewModel {
       void Function(DeviceInteractionViewModel$Change change) mutator) {
     final change = DeviceInteractionViewModel$Change._(
       this.deviceId,
+      this.connectableStatus,
       this.connectionStatus,
       this.deviceConnector,
       this.discoverServices,
@@ -38,6 +42,7 @@ abstract class $DeviceInteractionViewModel {
     mutator(change);
     return DeviceInteractionViewModel(
       deviceId: change.deviceId,
+      connectableStatus: change.connectableStatus,
       connectionStatus: change.connectionStatus,
       deviceConnector: change.deviceConnector,
       discoverServices: change.discoverServices,
@@ -46,7 +51,7 @@ abstract class $DeviceInteractionViewModel {
 
   @override
   String toString() =>
-      "DeviceInteractionViewModel(deviceId: $deviceId, connectionStatus: $connectionStatus, deviceConnector: $deviceConnector, discoverServices: $discoverServices)";
+      "DeviceInteractionViewModel(deviceId: $deviceId, connectableStatus: $connectableStatus, connectionStatus: $connectionStatus, deviceConnector: $deviceConnector, discoverServices: $discoverServices)";
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
@@ -54,6 +59,7 @@ abstract class $DeviceInteractionViewModel {
       other is DeviceInteractionViewModel &&
       other.runtimeType == runtimeType &&
       deviceId == other.deviceId &&
+      connectableStatus == other.connectableStatus &&
       connectionStatus == other.connectionStatus &&
       deviceConnector == other.deviceConnector &&
       const Ignore().equals(discoverServices, other.discoverServices);
@@ -63,6 +69,7 @@ abstract class $DeviceInteractionViewModel {
   int get hashCode {
     var result = 17;
     result = 37 * result + deviceId.hashCode;
+    result = 37 * result + connectableStatus.hashCode;
     result = 37 * result + connectionStatus.hashCode;
     result = 37 * result + deviceConnector.hashCode;
     result = 37 * result + const Ignore().hash(discoverServices);
@@ -73,15 +80,17 @@ abstract class $DeviceInteractionViewModel {
 class DeviceInteractionViewModel$Change {
   DeviceInteractionViewModel$Change._(
     this.deviceId,
+    this.connectableStatus,
     this.connectionStatus,
     this.deviceConnector,
     this.discoverServices,
   );
 
   String deviceId;
+  Connectable connectableStatus;
   DeviceConnectionState connectionStatus;
   BleDeviceConnector deviceConnector;
-  Future<List<DiscoveredService>> Function() discoverServices;
+  Future<List<Service>> Function() discoverServices;
 }
 
 // ignore: avoid_classes_with_only_static_members
@@ -90,6 +99,15 @@ class DeviceInteractionViewModel$ {
     (deviceIdContainer) => deviceIdContainer.deviceId,
     (deviceIdContainer, deviceId) =>
         deviceIdContainer.copyWith(deviceId: deviceId),
+  );
+
+  static final connectableStatus =
+      Lens<DeviceInteractionViewModel, Connectable>(
+    (connectableStatusContainer) =>
+        connectableStatusContainer.connectableStatus,
+    (connectableStatusContainer, connectableStatus) =>
+        connectableStatusContainer.copyWith(
+            connectableStatus: connectableStatus),
   );
 
   static final connectionStatus =
@@ -107,7 +125,7 @@ class DeviceInteractionViewModel$ {
   );
 
   static final discoverServices = Lens<DeviceInteractionViewModel,
-      Future<List<DiscoveredService>> Function()>(
+      Future<List<Service>> Function()>(
     (discoverServicesContainer) => discoverServicesContainer.discoverServices,
     (discoverServicesContainer, discoverServices) =>
         discoverServicesContainer.copyWith(discoverServices: discoverServices),
