@@ -1,29 +1,32 @@
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:meta/meta.dart';
+import 'package:reactive_ble_platform_interface/reactive_ble_platform_interface.dart';
 
 abstract class ConnectedDeviceOperation {
   Stream<CharacteristicValue> get characteristicValueStream;
 
-  Future<List<int>> readCharacteristic(QualifiedCharacteristic characteristic);
+  Future<List<int>> readCharacteristic(CharacteristicInstance characteristic);
 
   Future<void> writeCharacteristicWithResponse(
-    QualifiedCharacteristic characteristic, {
+    CharacteristicInstance characteristic, {
     required List<int> value,
   });
 
   Future<void> writeCharacteristicWithoutResponse(
-    QualifiedCharacteristic characteristic, {
+    CharacteristicInstance characteristic, {
     required List<int> value,
   });
 
   Stream<List<int>> subscribeToCharacteristic(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
     Future<void> isDisconnected,
   );
 
   Future<int> requestMtu(String deviceId, int mtu);
 
   Future<List<DiscoveredService>> discoverServices(String deviceId);
+
+  Future<List<DiscoveredService>> getDiscoverServices(String deviceId);
 
   Future<void> requestConnectionPriority(
       String deviceId, ConnectionPriority priority);
@@ -40,7 +43,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
       _blePlatform.charValueUpdateStream;
 
   @override
-  Future<List<int>> readCharacteristic(QualifiedCharacteristic characteristic) {
+  Future<List<int>> readCharacteristic(CharacteristicInstance characteristic) {
     final specificCharacteristicValueStream = characteristicValueStream
         .where((update) => update.characteristic == characteristic)
         .map((update) => update.result.dematerialize());
@@ -54,7 +57,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
 
   @override
   Future<void> writeCharacteristicWithResponse(
-    QualifiedCharacteristic characteristic, {
+    CharacteristicInstance characteristic, {
     required List<int> value,
   }) async =>
       _blePlatform
@@ -63,7 +66,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
 
   @override
   Future<void> writeCharacteristicWithoutResponse(
-    QualifiedCharacteristic characteristic, {
+    CharacteristicInstance characteristic, {
     required List<int> value,
   }) async =>
       _blePlatform
@@ -72,7 +75,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
 
   @override
   Stream<List<int>> subscribeToCharacteristic(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
     Future<void> isDisconnected,
   ) {
     final specificCharacteristicValueStream = characteristicValueStream
@@ -102,6 +105,10 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   @override
   Future<List<DiscoveredService>> discoverServices(String deviceId) =>
       _blePlatform.discoverServices(deviceId);
+
+  @override
+  Future<List<DiscoveredService>> getDiscoverServices(String deviceId) =>
+      _blePlatform.getDiscoverServices(deviceId);
 
   @override
   Future<void> requestConnectionPriority(
