@@ -13,25 +13,35 @@ import java.util.UUID
 @Suppress("TooManyFunctions")
 interface BleClient {
 
-    val connectionUpdateSubject: BehaviorSubject<com.signify.hue.flutterreactiveble.ble.ConnectionUpdate>
+    val connectionUpdateSubject: BehaviorSubject<ConnectionUpdate>
 
     fun initializeClient()
-    fun scanForDevices(services: List<ParcelUuid>, scanMode: ScanMode, requireLocationServicesEnabled: Boolean): Observable<com.signify.hue.flutterreactiveble.ble.ScanInfo>
+    fun scanForDevices(services: List<ParcelUuid>, scanMode: ScanMode, requireLocationServicesEnabled: Boolean): Observable<ScanInfo>
     fun connectToDevice(deviceId: String, timeout: Duration)
     fun disconnectDevice(deviceId: String)
     fun disconnectAllDevices()
     fun discoverServices(deviceId: String): Single<RxBleDeviceServices>
     fun clearGattCache(deviceId: String): Completable
-    fun readCharacteristic(deviceId: String, characteristic: UUID): Single<CharOperationResult>
-    fun setupNotification(deviceId: String, characteristic: UUID): Observable<ByteArray>
+    fun readCharacteristic(
+        deviceId: String,
+        characteristicId: UUID,
+        characteristicInstanceId: Int
+    ): Single<CharOperationResult>
+    fun setupNotification(
+        deviceId: String,
+        characteristicId: UUID,
+        characteristicInstanceId: Int
+    ): Observable<ByteArray>
     fun writeCharacteristicWithResponse(
         deviceId: String,
-        characteristic: UUID,
+        characteristicId: UUID,
+        characteristicInstanceId: Int,
         value: ByteArray
     ): Single<CharOperationResult>
     fun writeCharacteristicWithoutResponse(
         deviceId: String,
-        characteristic: UUID,
+        characteristicId: UUID,
+        characteristicInstanceId: Int,
         value: ByteArray
     ): Single<CharOperationResult>
     fun negotiateMtuSize(deviceId: String, size: Int): Single<MtuNegotiateResult>
