@@ -1,5 +1,6 @@
 package com.signify.hue.flutterreactiveble.channelhandlers
 
+import com.polidea.rxandroidble2.exceptions.BleDisconnectedException
 import com.signify.hue.flutterreactiveble.ProtobufModel as pb
 import com.signify.hue.flutterreactiveble.converters.ProtobufMessageConverter
 import com.signify.hue.flutterreactiveble.converters.UuidConverter
@@ -31,12 +32,23 @@ class CharNotificationHandler(private val bleClient: com.signify.hue.flutterreac
     fun subscribeToNotifications(request: pb.NotifyCharacteristicRequest) {
         val charUuid = uuidConverter
             .uuidFromByteArray(request.characteristic.characteristicUuid.data.toByteArray())
-        val subscription = bleClient.setupNotification(request.characteristic.deviceId, charUuid)
+        val subscription = bleClient.setupNotification(
+                request.characteristic.deviceId,
+                charUuid,
+                request.characteristic.characteristicInstanceId.toInt()
+        )
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({ value ->
                 handleNotificationValue(request.characteristic, value)
             }, {
-                handleNotificationError(request.characteristic, it)
+                when (it) {
+                    is BleDisconnectedException -> {
+                        subscriptionMap.remove(request.characteristic)?.dispose()
+                    }
+                    else -> {
+                        handleNotificationError(request.characteristic, it)
+                    }
+                }
             })
         subscriptionMap[request.characteristic] = subscription
     }
