@@ -36,6 +36,9 @@ class ProtobufMessageConverter {
                     .setId(scanInfo.deviceId)
                     .setName(scanInfo.name)
                     .setRssi(scanInfo.rssi)
+                    .setIsConnectable(pb.IsConnectable.newBuilder()
+                            .setCode(scanInfo.connectable.code)
+                            .build())
                     .addAllServiceData(createServiceDataEntry(scanInfo.serviceData))
                     .addAllServiceUuids(createServiceUuids(scanInfo.serviceUuids))
                     .setManufacturerData(ByteString.copyFrom(scanInfo.manufacturerData))
@@ -93,13 +96,13 @@ class ProtobufMessageConverter {
             request: pb.CharacteristicAddress,
             error: String?
     ): pb.CharacteristicValueInfo {
-        val characteristicAdress = createCharacteristicAddress(request)
+        val characteristicAddress = createCharacteristicAddress(request)
         val failure = pb.GenericFailure.newBuilder()
                 .setCode(CharacteristicErrorType.UNKNOWN.code)
                 .setMessage(error ?: "Unknown error")
 
         return pb.CharacteristicValueInfo.newBuilder()
-                .setCharacteristic(characteristicAdress)
+                .setCharacteristic(characteristicAddress)
                 .setFailure(failure)
                 .build()
     }
@@ -176,6 +179,7 @@ class ProtobufMessageConverter {
     private fun fromBluetoothGattService(gattService: BluetoothGattService): pb.DiscoveredService {
         return pb.DiscoveredService.newBuilder()
                 .setServiceUuid(createUuidFromParcelUuid(gattService.uuid))
+                .setServiceInstanceId(gattService.instanceId.toString())
                 .addAllCharacteristicUuids(gattService.characteristics.map { createUuidFromParcelUuid(it.uuid) })
                 .addAllCharacteristics(gattService.characteristics.map {
                     val prop = it.properties
@@ -187,6 +191,7 @@ class ProtobufMessageConverter {
 
                     pb.DiscoveredCharacteristic.newBuilder()
                             .setCharacteristicId(createUuidFromParcelUuid(it.uuid))
+                            .setCharacteristicInstanceId(it.instanceId.toString())
                             .setServiceId(createUuidFromParcelUuid(it.service.uuid))
                             .setIsReadable(readable)
                             .setIsWritableWithResponse(write)
@@ -215,11 +220,12 @@ class ProtobufMessageConverter {
         return pb.CharacteristicAddress.newBuilder()
                 .setDeviceId(request.deviceId)
                 .setServiceUuid(request.serviceUuid)
+                .setServiceInstanceId(request.serviceInstanceId)
+                .setCharacteristicInstanceId(request.characteristicInstanceId)
                 .setCharacteristicUuid(request.characteristicUuid)
     }
 
     private fun createServiceDataEntry(serviceData: Map<UUID, ByteArray>): List<pb.ServiceDataEntry> {
-
         val serviceDataEntries = mutableListOf<pb.ServiceDataEntry>()
 
         // Needed ugly for-loop because we support API23 that does not support kotlin foreach
