@@ -13,8 +13,8 @@ struct CharacteristicWriteTaskController: PeripheralTaskController {
     func start(peripheral: CBPeripheral) -> SubjectTask {
         guard
             peripheral.state == .connected,
-            let service = peripheral.services?.first(where: { $0.uuid == task.key.serviceID }),
-            let characteristic = service.characteristics?.first(where: { $0.uuid == task.key.id }),
+            let service = peripheral.services?.filter({ $0.uuid == task.key.serviceID })[Int(task.key.serviceInstanceID) ?? 0],
+            let characteristic = service.characteristics?.filter({ $0.uuid == task.key.id })[Int(task.key.instanceID) ?? 0],
             characteristic.properties.contains(.write)
         else {
             return task.with(state: task.state.finished(PluginError.internalInconcictency(details: nil)))
