@@ -62,6 +62,7 @@ class ProtobufConverterImpl implements ProtobufConverter {
           serviceUuids: serviceUuids,
           manufacturerData: Uint8List.fromList(message.manufacturerData),
           rssi: message.rssi,
+          connectable: _connectableFrom(message.isConnectable),
         ),
         failure: genericFailureFrom(
             hasFailure: message.hasFailure(),
@@ -162,10 +163,12 @@ class ProtobufConverterImpl implements ProtobufConverter {
   int mtuSizeFrom(List<int> data) =>
       pb.NegotiateMtuInfo.fromBuffer(data).mtuSize;
 
-  QualifiedCharacteristic qualifiedCharacteristicFrom(
+  CharacteristicInstance qualifiedCharacteristicFrom(
           pb.CharacteristicAddress message) =>
-      QualifiedCharacteristic(
+      CharacteristicInstance(
+        characteristicInstanceId: message.characteristicInstanceId,
         characteristicId: Uuid(message.characteristicUuid.data),
+        serviceInstanceId: message.serviceInstanceId,
         serviceId: Uuid(message.serviceUuid.data),
         deviceId: message.deviceId,
       );
@@ -198,6 +201,7 @@ class ProtobufConverterImpl implements ProtobufConverter {
   DiscoveredService _convertService(pb.DiscoveredService service) =>
       DiscoveredService(
         serviceId: Uuid(service.serviceUuid.data),
+        serviceInstanceId: service.serviceInstanceId,
         characteristicIds: service.characteristicUuids
             .map((c) => Uuid(c.data))
             .toList(growable: false),
@@ -205,6 +209,7 @@ class ProtobufConverterImpl implements ProtobufConverter {
             .map((c) => DiscoveredCharacteristic(
                 characteristicId: Uuid(c.characteristicId.data),
                 serviceId: Uuid(c.serviceId.data),
+                characteristicInstanceId: c.characteristicInstanceId,
                 isReadable: c.isReadable,
                 isWritableWithResponse: c.isWritableWithResponse,
                 isWritableWithoutResponse: c.isWritableWithoutResponse,
@@ -224,6 +229,19 @@ class ProtobufConverterImpl implements ProtobufConverter {
       failure != null
           ? Result<Value, Failure>.failure(failure)
           : Result.success(getValue());
+
+  Connectable _connectableFrom(pb.IsConnectable status) {
+    switch (status.code) {
+      case 0:
+        return Connectable.unknown;
+      case 1:
+        return Connectable.unavailable;
+      case 2:
+        return Connectable.available;
+    }
+
+    return Connectable.unknown;
+  }
 }
 
 class _InvalidConnectionState extends Error {
