@@ -155,7 +155,7 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
   }
 
   @override
-  Stream<void> readCharacteristic(QualifiedCharacteristic characteristic) {
+  Stream<void> readCharacteristic(CharacteristicInstance characteristic) {
     _logger?.log(
       'Read characteristic: $characteristic',
     );
@@ -171,7 +171,7 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
 
   @override
   Future<WriteCharacteristicInfo> writeCharacteristicWithResponse(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
     List<int> value,
   ) async {
     _logger?.log('Write with response to $characteristic, value: $value');
@@ -186,7 +186,7 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
 
   @override
   Future<WriteCharacteristicInfo> writeCharacteristicWithoutResponse(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
     List<int> value,
   ) async {
     _logger?.log(
@@ -204,7 +204,7 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
 
   @override
   Stream<void> subscribeToNotifications(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   ) {
     _logger?.log('Start subscribing to notifications for $characteristic');
     return _bleMethodChannel
@@ -219,7 +219,7 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
 
   @override
   Future<void> stopSubscribingToNotifications(
-    QualifiedCharacteristic characteristic,
+    CharacteristicInstance characteristic,
   ) {
     _logger?.log('Stop subscribing to notifications for $characteristic');
     return _bleMethodChannel
@@ -283,6 +283,19 @@ class ReactiveBleMobilePlatform extends ReactiveBlePlatform {
     return _bleMethodChannel
         .invokeMethod<List<int>>(
           'discoverServices',
+          _argsToProtobufConverter
+              .createDiscoverServicesRequest(deviceId)
+              .writeToBuffer(),
+        )
+        .then((data) => _protobufConverter.discoveredServicesFrom(data!));
+  }
+
+  @override
+  Future<List<DiscoveredService>> getDiscoverServices(String deviceId) async {
+    _logger?.log('Get discovered services for device: $deviceId');
+    return _bleMethodChannel
+        .invokeMethod<List<int>>(
+          'getDiscoveredServices',
           _argsToProtobufConverter
               .createDiscoverServicesRequest(deviceId)
               .writeToBuffer(),
