@@ -1,10 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:reactive_ble_mobile/src/converter/args_to_protubuf_converter.dart';
 import 'package:reactive_ble_mobile/src/generated/bledata.pb.dart' as pb;
-import 'package:reactive_ble_platform_interface/src/model/connection_priority.dart';
-import 'package:reactive_ble_platform_interface/src/model/qualified_characteristic.dart';
-import 'package:reactive_ble_platform_interface/src/model/scan_mode.dart';
-import 'package:reactive_ble_platform_interface/src/model/uuid.dart';
+import 'package:reactive_ble_platform_interface/reactive_ble_platform_interface.dart';
 
 void main() {
   group('$ArgsToProtobufConverter', () {
@@ -92,28 +89,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createReadCharacteristicRequest(characteristic);
       });
 
-      test('It converts device Id ', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
@@ -122,30 +117,28 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       const value = [0, 1];
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createWriteCharacteristicRequest(characteristic, value);
       });
 
-      test('It converts device Id ', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
 
       test('It converts value', () {
@@ -158,28 +151,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
 
         result = _sut.createNotifyCharacteristicRequest(characteristic);
       });
 
-      test('It converts device Id ', () {
+      test('It converts device ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts service Uuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts char Uuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
@@ -317,28 +308,26 @@ void main() {
       const deviceId = '123';
       final serviceUuid = Uuid.parse('FEFF');
       final charUuid = Uuid.parse('FEEF');
-      QualifiedCharacteristic characteristic;
+      CharacteristicInstance characteristic;
       late pb.NotifyNoMoreCharacteristicRequest result;
 
       setUp(() {
-        characteristic = QualifiedCharacteristic(
+        characteristic = CharacteristicInstance(
           characteristicId: charUuid,
+          characteristicInstanceId: "101",
           serviceId: serviceUuid,
+          serviceInstanceId: "11",
           deviceId: deviceId,
         );
         result = _sut.createNotifyNoMoreCharacteristicRequest(characteristic);
       });
 
-      test('It converts deviceId', () {
+      test('It converts the ids', () {
         expect(result.characteristic.deviceId, deviceId);
-      });
-
-      test('It converts serviceUuid', () {
         expect(result.characteristic.serviceUuid.data, [254, 255]);
-      });
-
-      test('It converts characteristicUuid', () {
+        expect(result.characteristic.serviceInstanceId, "11");
         expect(result.characteristic.characteristicUuid.data, [254, 239]);
+        expect(result.characteristic.characteristicInstanceId, "101");
       });
     });
 
