@@ -71,6 +71,7 @@ abstract class $DiscoveredDevice {
   List<Uuid> get serviceUuids;
   Uint8List get manufacturerData;
   int get rssi;
+  Connectable get connectable;
 
   DiscoveredDevice copyWith({
     String? id,
@@ -79,6 +80,7 @@ abstract class $DiscoveredDevice {
     List<Uuid>? serviceUuids,
     Uint8List? manufacturerData,
     int? rssi,
+    Connectable? connectable,
   }) =>
       DiscoveredDevice(
         id: id ?? this.id,
@@ -87,6 +89,7 @@ abstract class $DiscoveredDevice {
         serviceUuids: serviceUuids ?? this.serviceUuids,
         manufacturerData: manufacturerData ?? this.manufacturerData,
         rssi: rssi ?? this.rssi,
+        connectable: connectable ?? this.connectable,
       );
 
   DiscoveredDevice copyUsing(
@@ -98,6 +101,7 @@ abstract class $DiscoveredDevice {
       this.serviceUuids,
       this.manufacturerData,
       this.rssi,
+      this.connectable,
     );
     mutator(change);
     return DiscoveredDevice(
@@ -107,12 +111,13 @@ abstract class $DiscoveredDevice {
       serviceUuids: change.serviceUuids,
       manufacturerData: change.manufacturerData,
       rssi: change.rssi,
+      connectable: change.connectable,
     );
   }
 
   @override
   String toString() =>
-      "DiscoveredDevice(id: $id, name: $name, serviceData: $serviceData, serviceUuids: $serviceUuids, manufacturerData: $manufacturerData, rssi: $rssi)";
+      "DiscoveredDevice(id: $id, name: $name, serviceData: $serviceData, serviceUuids: $serviceUuids, manufacturerData: $manufacturerData, rssi: $rssi, connectable: $connectable)";
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
@@ -125,7 +130,8 @@ abstract class $DiscoveredDevice {
       const DeepCollectionEquality().equals(serviceUuids, other.serviceUuids) &&
       const DeepCollectionEquality()
           .equals(manufacturerData, other.manufacturerData) &&
-      rssi == other.rssi;
+      rssi == other.rssi &&
+      connectable == other.connectable;
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
@@ -138,6 +144,7 @@ abstract class $DiscoveredDevice {
     result =
         37 * result + const DeepCollectionEquality().hash(manufacturerData);
     result = 37 * result + rssi.hashCode;
+    result = 37 * result + connectable.hashCode;
     return result;
   }
 }
@@ -150,6 +157,7 @@ class DiscoveredDevice$Change {
     this.serviceUuids,
     this.manufacturerData,
     this.rssi,
+    this.connectable,
   );
 
   String id;
@@ -158,6 +166,7 @@ class DiscoveredDevice$Change {
   List<Uuid> serviceUuids;
   Uint8List manufacturerData;
   int rssi;
+  Connectable connectable;
 }
 
 // ignore: avoid_classes_with_only_static_members
@@ -193,5 +202,11 @@ class DiscoveredDevice$ {
   static final rssi = Lens<DiscoveredDevice, int>(
     (rssiContainer) => rssiContainer.rssi,
     (rssiContainer, rssi) => rssiContainer.copyWith(rssi: rssi),
+  );
+
+  static final connectable = Lens<DiscoveredDevice, Connectable>(
+    (connectableContainer) => connectableContainer.connectable,
+    (connectableContainer, connectable) =>
+        connectableContainer.copyWith(connectable: connectable),
   );
 }
