@@ -2,7 +2,7 @@ import 'package:meta/meta.dart';
 
 import 'uuid.dart';
 
-/// Specific BLE characteristic for a BLE device characterised by [deviceId], [serviceId] and
+/// Specific BLE characteristic for a BLE device characterized by [deviceId], [serviceId] and
 /// [characteristicId].
 @immutable
 class QualifiedCharacteristic {
